egistec: Rename Nile to legacy and Ganges to current.
The "API" for the new TZ QSEE app found on the Ganges platform is
applicable to updated Nile devices and (with slight alteration) used on
Kumano too. Hence it is (without a proper version number) more obvious
to rename both implementations to legacy and current, which is more
generic.
